requirements.txt
gtts==2.5.0
python-docx==1.1.2
